requirements.txt
mcp-plugin-tools
openai
python-dotenv
pydantic
tavily-python
